refactor: extract comment box to its own component
