+++ ui/components/prsidebar/prsidebar.go
@@ -9,6 +9,7 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 	"github.com/dlvhdr/gh-dash/data"
+	"github.com/dlvhdr/gh-dash/ui/components/commentbox"
 	"github.com/dlvhdr/gh-dash/ui/components/pr"
 	"github.com/dlvhdr/gh-dash/ui/context"
 	"github.com/dlvhdr/gh-dash/ui/markdown"
@@ -21,42 +22,17 @@ type Model struct {
 	pr           *pr.PullRequest
 	width        int
 	isCommenting bool
-	textArea     textarea.Model
+	commentBox   commentbox.Model
 	commentHelp  help.Model
 }
 
 func NewModel() Model {
-	ta := textarea.New()
-	ta.ShowLineNumbers = true
-	ta.Prompt = ""
-	ta.FocusedStyle.Base = lipgloss.NewStyle()
-	ta.FocusedStyle.CursorLine = lipgloss.NewStyle().
-		Background(styles.DefaultTheme.FaintBorder).
-		Foreground(styles.DefaultTheme.PrimaryText)
-	ta.FocusedStyle.LineNumber = lipgloss.NewStyle().Foreground(styles.DefaultTheme.FaintText)
-	ta.FocusedStyle.CursorLineNumber = lipgloss.NewStyle().Foreground(styles.DefaultTheme.SecondaryText)
-	ta.FocusedStyle.Placeholder = lipgloss.NewStyle().Foreground(styles.DefaultTheme.FaintText)
-	ta.FocusedStyle.Text = lipgloss.NewStyle().Foreground(styles.DefaultTheme.PrimaryText)
-	ta.FocusedStyle.EndOfBuffer = lipgloss.NewStyle().Foreground(styles.DefaultTheme.FaintText)
-	ta.Focus()
-
-	helpTextStyle := lipgloss.NewStyle().Foreground(styles.DefaultTheme.SecondaryText)
-	h := help.NewModel()
-	h.Styles = help.Styles{
-		ShortDesc:      helpTextStyle.Copy().Foreground(styles.DefaultTheme.FaintText),
-		FullDesc:       helpTextStyle.Copy(),
-		ShortSeparator: helpTextStyle.Copy().Foreground(styles.DefaultTheme.SecondaryBorder),
-		FullSeparator:  helpTextStyle.Copy(),
-		FullKey:        helpTextStyle.Copy().Foreground(styles.DefaultTheme.PrimaryText),
-		ShortKey:       helpTextStyle.Copy(),
-		Ellipsis:       helpTextStyle.Copy(),
-	}
+	commentBox := commentbox.NewModel()
 
 	return Model{
 		pr:           nil,
 		isCommenting: false,
-		textArea:     ta,
-		commentHelp:  h,
+		commentBox:   commentBox,
 	}
 }
 
@@ -77,21 +53,21 @@ func (m Model) Update(msg tea.Msg) (Model, tea.Cmd) {
 		switch msg.Type {
 
 		case tea.KeyCtrlD:
-			if len(strings.Trim(m.textArea.Value(), " ")) != 0 {
-				cmd = m.comment(m.textArea.Value())
+			if len(strings.Trim(m.commentBox.Value(), " ")) != 0 {
+				cmd = m.comment(m.commentBox.Value())
 			}
-			m.textArea.Blur()
+			m.commentBox.Blur()
 			m.isCommenting = false
 			return m, cmd
 
 		case tea.KeyEsc, tea.KeyCtrlC:
-			m.textArea.Blur()
+			m.commentBox.Blur()
 			m.isCommenting = false
 			return m, nil
 		}
 	}
 
-	m.textArea, taCmd = m.textArea.Update(msg)
+	m.commentBox, taCmd = m.commentBox.Update(msg)
 	cmds = append(cmds, cmd, taCmd)
 	return m, tea.Batch(cmds...)
 }
@@ -111,7 +87,7 @@ func (m Model) View() string {
 	s.WriteString(m.renderActivity())
 
 	if m.isCommenting {
-		s.WriteString(m.renderCommentBox())
+		s.WriteString(m.commentBox.View())
 	}
 
 	return s.String()
@@ -230,7 +206,7 @@ func (m *Model) SetRow(data *data.PullRequestData) {
 
 func (m *Model) SetWidth(width int) {
 	m.width = width
-	m.textArea.SetWidth(width)
+	m.commentBox.SetWidth(width)
 }
 
 func (m *Model) GetIsCommenting() bool {
@@ -243,12 +219,12 @@ func (m *Model) UpdateProgramContext(ctx *context.ProgramContext) {
 
 func (m *Model) SetIsCommenting(isCommenting bool) tea.Cmd {
 	if m.isCommenting == false && isCommenting == true {
-		m.textArea.Reset()
+		m.commentBox.Reset()
 	}
 	m.isCommenting = isCommenting
 
 	if isCommenting == true {
-		return tea.Sequentially(textarea.Blink, m.textArea.Focus())
+		return tea.Sequentially(textarea.Blink, m.commentBox.Focus())
 	}
 	return nil
 }
